Cleaned up the messes I had made in the client classes. Fixed a potential bug with the split char in GWTClient still splitting on "+" instead of "*".
Now that I think of it, I dunno why I changed that. Oh well.

Added a List<String> to help out people who like polling for messages
rather than the PNCME.
Added support for clients in the PNCME since I had completely forgotten
that clients really don't need to keep track of a clientID when dealing
with a message.
Added support for the PNCME in the client classes.